+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.6.8`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.0`
 
 ## [0.5.1] - 2024-03-29