+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `gradle -> 8.8`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.0`
 
 ## [0.5.1] - 2024-03-29