+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `gradle -> 8.8`
+- `gradle -> 8.9`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.0`
 
 ## [0.5.1] - 2024-03-29