+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.1`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.7.2`
 - `gradle -> 8.9`
 
 ## [0.5.1] - 2024-03-29